refactor(paintwidget): split PaintWidget in 3 files (index, state and behavior)
This commit refactors the PaintWidget to separate its factory (in index.js), its state (in state.js)
and its behavior (in behavior.js)
